requirements.txt
Flask==2.3.3
flask-cors==4.0.0
gunicorn==20.1.0
boto3==1.34.38
python-dotenv==1.0.1
Pillow==10.2.0
opencv-python==4.9.0.80
numpy==1.26.4
deepface==0.0.79
